Add functions to find regions in an FDT
Given a set of nodes and properties, find the regions of the device tree
which describe those parts.

Put this feature in a new libfdt_extra library.

A test is provided which builds a tree while tracking where the regions
should be, then calls fdt_first/next_region() to make sure that it agrees.

Further tests will come as part of fdtgrep.

Signed-off-by: Simon Glass <sjg@chromium.org>